requirements.txt
slack-sdk>=3.23.0
flask>=3.0.0
pandas>=2.0.0
plotly>=5.17.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
